moved all tests to the test folder
